Add userdebug exceptions for vcd and dmd ports in ListeningPortsTest
The ListeningPortsTest will report a failure if an unexpected system
process (uid < 10000) is listening on a port and can be accessed by
other apps on the device. The dmd and vcd processes are expected
to be listening on the loopback interface to ports 50002 and 60002
on userdebug builds, so this commit adds exceptions to not report
these ports if the device is running a userdebug or eng build.

Fixes: 281490175
Test: atest ListeningPortsTest
Change-Id: I625e310cb34378e0dd97cdb50a3489689c086664