Catch state exception in instrumentation close
If we receive an IllegalStateException in the close method of the
observer, that means we are attempting to close prior to the framework
instrumentation class attaching.

This may happen for quick tests which don't actually utilize the
instrumentation.

We can safely suppress this exception, since if the instrumentation has
not attached, there is nothing to clean up.

Bug: 282156588
Test: CtsVoiceInteractionTestCases
Test: CtsSoundTriggerTestCases
Change-Id: I341c9c93dd3ab53462a96af49cb2b00097e95af4