disconnect and destroy the created call to prevent stuck calls
A set of test failures came for the SelfManagedConnectionServiceTest.
Upon inspecting the culprit CLs, there was only 2 telecom changes that
merged within the breakage window.

The stack trace says a call is not permitted for the failing tests. This
is an indication that the device could be in a bad state (e.g. max
ongoing calls reached, etc.).

One of the CLs in the window creates calls but does not destroy them.
That is most likely the issue causing these test failures.  At the very
least, the test should be disconnect and destroying the call it created.
At best, this will clear up the test failures.

Fixes: 282158278
Test: refactor testMidCallComponentEnablementWithNoneAvailableAtStart
Change-Id: I735628aa782948299ce420517007f3ff8164660b