[ITS][SF]: initialize Arduino before rotating DUT
Details: created a separate method to initialize the Arduino and raise
an AssertionError in the parent process if initialization fails. This is
to ensure that the Arduino is always initialized before sending the
rotation command. After initialization, a child process can be created
to do the rotation while taking a recording. Also, increased timeout to
have other tests wait for sensor_fusion to finish.

Bug: 246684281
Test: tested remotely
Change-Id: Ib830d3efb73f142ae424afa7b0fd74d89c788d87