Fix failing UserVisibilityTest#testIsUserVisible_backgroundUser in HSUM.
In HSUM, DeviceState#switchFromUser failed to switch from the current
user, thus user 10 stayed as current foreground user for the test.
By using ensureHasAdditionalUser, we are able to create a new secondary
user and switch to it. This change will not affect non-HSUM cases.

Bug: 283274303

Test: atest UserVisibilityTest (for both HSUM and non-HSUM targets)

Ignore-AOSP-First: the fix needed only for a test added in U.

Change-Id: Ib87eb32efcbca0df9af6153e390a3a2d534ee55a