Allow requesting one burst size data after stopping a legacy stream.
When stopping the legacy stream, the callback thread is not joined
immediately. Instead, a flag is set to suspend the callback thread,
there will be no callback when entering a new thread loop after setting
the flag. In that case, the test should allow requesting one burst size
data from callback after stopping a legacy stream.

Bug: 282924185
Test: atest AAudioTests
Test: atest AAudioTests --test-filter="android.nativemedia.aaudio.AAudioTests#SPM_AAudioOutputStreamCallbackTest_testPlayback_SHARED__67__DEFAULT__MMAPOK__UNSP" --iteration 100
Change-Id: I262218a723069f7a7fb9bd7c0e73dbdb41c70ece