Fix ActivityLifecycleTopResumedStateTests for legacy transition devices
The trampoline activity was not in the animating state in legacy
transition, while the shell transition collects all WindowContainer
changes and also evaluated the trampoline activity as animating.
Since the activity won't be stopped if the invisible activity is
currently animating, the stop lifecycle of the test activities can be
different between legacy and shell transition.

Fix: 275295202
Fix: 281765440
Test: adb shell setprop persist.wm.debug.shell_transit 0 && atest \
    android.server.wm.lifecycle.ActivityLifecycleTopResumedStateTests
Change-Id: I5e9a0fcbf040d4b8d91a26e1c91b1b12505eabbc