Fix DockConfigChangeTests by rotating device before docking
Devices on CTS seem to always start in portrait mode at 0 degrees of
rotation. Since docking rotates the device to 270 degreees, that causes
an extra config change event, which fails the test.

In development, I've always used the device horizontally, which caused
the test to pass locally.

Bug: 283917852
Test: atest DockConfigChangeTests
(tested with device starting from portrait and landscape)

Change-Id: Ic58d254baa4c45a620c037d0d55bf46ed26faa61