Skip AOHD compat change tests on user builds
The compat change modification APIs we use relies on userdebug for tests
which override a change when targetSdk is the same as EnabledSince.

As a workaround until this limitation is removed or this test is pulled
into a separate apk, disable the test on non-debuggable (user) builds.

Add a test rule which skips tests conditionally, since we must
short-circuit before the problematic rule is evaluated.

Fixes: 283378311
Test: AlwaysOnHotwordDetectorNoDspTest on GSI/user and cf/userdebug
Change-Id: I918686dcf236fcbab126f79ca4704bfb850dc092