Fixed inset offset bug for a test and better logging for tests
AttachedSurfaceControlTest doesn't respect inset bounds when layed out.
The test would offset the bounds to check by the inset of the main
window which was inaccurate. Instead, pass in no insets when checking
screenshot for this test.

Also changed logging info so the caller can specify a region to log when
pixels mismatch since often it only cares about a specific area of
matching pixels

Test: AttachedSurfaceControlTest
Test: SurfaceViewSurfaceValidatorTest
Bug: 284247638
Change-Id: Ib86f2cbc5e556e683cc8a54fd45b8050bb2e4b3b