Fix ClipboardAccessTest NPE in HALF_FOLDED mode
A bad fix skips tests in HALF_FOLDED mode this will cause @Before in
ClipboardAccessTest will not be executed. The mClipboardManager is
set in @Before, the bad fix will cause the mClipboardManager not be
set. So NPE occurs in @After because test think the @Before will be
executed.

Because the same process should be the same ClipboardManager, we will
try to get ClipboardManager again instead of using mClipboardManager.

Bug: 284425542
Test: atest ClipboardAccessTest with fold device
Change-Id: I14168785e06bf87363e9d609ca57508545edbc36