Invoke endRearDisplaySession with ControlDeviceState Permission
DeviceStateManagerService's isForegroundApp check uses
ActivityManagerService#getRunningAppProcesses to check if the caller
is IMPORTANCE_FOREGROUND. This usually happens quickly, but sometimes
is not fast enough in an environment such as CTS where the activity
launches and immediately invokes this code path. To de-flake this
test, we invoke endRearDisplaySession with the ControlDeviceState
permission - the same thing that has been done already for
startRearDisplaySession.

Bug: 281952366
Test: atest --iterations 20 android.server.wm.jetpack.area.ExtensionRearDisplayTest#testStartAndEndRearDisplaySession
Change-Id: Ib99d1efd1cdeb49dd1d5ee63315148619302e9b2