Remove unnecessary cts tag.
Some modules defined in the build file are only data dependencies of cts test modules. In this case, they will be packaged to corresponding test modules. For example, if module A is a data dependency of test module B, the module A can be found under android-cts/testcases/B/ directory. Therefore it is unnecesasry to tag data dependencies which will cause a duplicate folder, e.g. android-cts/testcases/A.

Change-Id: I16c1d2b735bf4b88be1272a5ecfaf74032f7dc8a
Test: m cts
Bug: 283199104