Moving testCrossUserMediaAccessInPublicSdCard to U+ only.
The original bug-fix for this issue was merged as aosp/2423165 and did not reach tm-qpr-dev. 
The test however, is mistakingly marked to run on T+, although the fix was only U+

Consequently, failures like https://android-build.corp.google.com/test_investigate/?invocationId=I74600010165411062&testResultId=TR83628696775259734 are seen when the test is attempted as a part of MediaProvider MTS (based out of udc-mainline-prod, including the test) on a T build, not containing the required changes.

To fix the same we move the test to target only U+.

Bug: 285554262
Change-Id: I888ec48e9cfc4a31d714c91de6127f3d5f5995ce
Test: atest CtsAppCloningPublicVolumeTest