Remove unnecessary cts tag.
Some modules defined in the build file are only data dependencies of cts test modules. In this case, they will be packaged to corresponding test modules. For example, if module A is a data dependency of test module B, the module A can be found under android-cts/testcases/B/ directory. Therefore it is unnecessary to tag data dependencies which will cause a duplicate folder, e.g. android-cts/testcases/A.

Change-Id: If98b266dfe295661664b5244b7fbb3c159147335
Test: m cts
Bug: 283199104