Make sure a display is tracked for accessibility
This CTS test is flaky because it's possible we register a proxy for a
display before that display has propagated to the accessibility
framework. If we first launch an activity and wait for it be be stable on
screen, then this display is guaranteed to be available before
continuinig the test.

Test: atest testRegisterAccessibilityProxy_withNonDeviceDisplay_throwsSecurityException 30x
Bug: 285243269
Change-Id: I3b5c83038457870a2139635484dfded68d6b76b7