Automotive Test Failure Fix
The CTS test was successful for all presubmit runs for our past check
in. However, a new bug arose that indicated these tests fail for
automotive UDC branch. Thus, the suspected culprit is that automotive
devices do not have the credential manager feature, so this adds a rule
to ensure the test must have said feature before running.

Bug: 286172710
Test: Will need to wait for presubmit tests for Automotive UDC
Change-Id: I64e5b85150f4ea7af982b681d54585a4e37a6249