MediaRecorderTest: Fixed the  android.media.recorder.cts.MediaRecorderTest#testRecordExceedFileSizeLimit fail
Since I91440cbfe54ade5128a280ed4cd1bcd248d06982 is expected to use the maximum resolution of the camera for testing, so the filesize is increased to adapt. However, the minimum resolution test is still used at present, which will cause the test time to be too long and time out. We recommend replacing CamcorderProfile.QUALITY_LOW with CamcorderProfile.QUALITY_HIGH.

Bug: 286718807
Test: CtsMediaRecorderTestCases android.media.recorder.cts.MediaRecorderTest#testRecordExceedFileSizeLimit

Change-Id: Ia6db0d3d606d8e1129bbbfbbdc9d3e74b3710a10