Fix CtsCarrierApiTestCases failure on OEM devices
This CL looses the condition on SIM state requirement for a corner case
on UDC based devices where SIM states is NOT unified. The same test case
will be updated to be much more restrict on VIC release when unified
SIM states are required and implemented.

Two devices from two OEMs reproduced the same failure when performing
Carrier API CTS case. When test SIM is powered down and waited to
power up to PASS_THROUGH mode, there is no deterministic approach to
detect, when device has RADIO HAL 1.6 and above.

The current ApduScriptUtil logic restricts too much (the SimAppState has
to be the same as targetSimState which is not always guaranteed for all
OEMs or even Pixels due to no explict requirement in this corner case).

Bug: 286324661
Test: atest CtsCarrierApiTestCases
Change-Id: Ia9c6b4410e604c1f4ab1c975f8465bac58795a7a