Refactor tests to reduce latency
Instead of starting both apps and sending pending intents between them
to orchestrate the flow, only start the FG app and do the orchestration
through calls to a test service that creates/sends the pending intents.

While this is a mitigation for 2 specific tests being flaky (to the
extend that it might reliably fail if the test system is slow) it can
also be a prototype how most tests in this class should be rewritten.

This only makes tests more reliable and there is no significant risk
that this might break anything.

Test: atest BackgroundActivityLaunchTest
Bug: 286917941
Change-Id: Ief0248c7f39dbc4d051e9ee45f1d5fab8433acfa