Set peek value after codec starts
MediaCodec.start() overwrites the peek state if it is set to off
before start is called, so the peek value needs to be set after
start is called. To allow setting peek off before start in next
release.

Set MediaCodec.Callback before codec configure, as indicated in
MediaCodec documentation

Additionally, sleep when AudioTrack.write does not write the full
buffer, because the track is full. Also need to run audio callback
in a separate thread to avoid blocking other operations.

MUST_SLEEP

Test: atest DecodeOnlyTest
Bug: 285365381
Change-Id: Ia6192fc3770754250dde89b071e028abff53b27b