Attempts fixing testChangeFontScaleNoRelaunch flakiness
For some reasons, the font scale is not updated sometimes after
setting the new font scale on cf. So, the test failed. The is
happening while updating the font scale from 1.0 to 1.15 based on
the local trials.

Reduce the number of font scale changes since it is not really
needed to make the font scale changes multiple times.

Bug: 285569662
Test: atest ConfigChangeTests on cf over 400 times
Change-Id: I6856a6d20f05b5db0ffc117db4304b78fbec1496