Don't always wait for permission consent in CapturedActivity
There are cases where the test running with CapturedActivity has the
manifest permission CAPTURE_VIDEO_OUTPUT and is running as signature. In
this case, MediaProjection will not show the consent dialog because it's
automatically granted. The test can't wait to the dismiss the consent
dialog since it won't be shown.

Skip waiting if the MediaProjection token is valid even if the consent
dialog was never shown.

Test: SurfaceViewSyncContinuousTest
Bug: 253218374
Change-Id: Ia4837c6c921019f7c9141570651405329fa2ca4a