Optimize testAeModeAndLock
To reduce test time for testAeModeAndLock, we should run the test with
smart combinations of exposure and sensitivity.

Instead of testing all combinations, we should test:
- minimum exposure in combination with all sensitivities,
- other exposures only with min and max sensitivities.

This maintains the test coverage while reducing the test time
drastically.
The test time for Oriole on the latest stable build (10337921) goes
down from ~4m14s to ~1m5s (~75% improvement).

Bug: 287260354
Test: atest CaptureRequestTest#testAeModeAndLock
Change-Id: I7ad5b53ca62baf16d4adf31370229a9302dcfe28