DO NOT MERGE: Move HorizontalScrollView to middle Fixes: 279501424
The HorizontalScrollView that was used for testing the stretch
was up against the side of the screen. Even though the drag
was from the center of the HorizontalScrollView, the View was
very small and the touch event was being treated as a system
gesture.

The HorizontalScrollView was moved to the center of the
screen to avoid the edges.

Test: HorizontalScrollView on raven
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:67f0f512c14eba596ca003d912be7ddafa06b730)
Merged-In: Ic690da8fce97af4f0829abec370cec48a17bad6e
Change-Id: Ic690da8fce97af4f0829abec370cec48a17bad6e