Skip VirtualSensor listener tests if there are no real events.
The device.poll() call in SensorService::threadLoop is blocking, so if
it doesn't return any events from the HAL, any virtual sensor events
will never reach the respective listeners.

This is a CTS fix to skip the relevant tests if there are no real
events coming from the default device accelerometer. Verified that
this fixes the issue by sleeping in AidlSensorHalWrapper::pollFmq
and ensuring that the tests are skipped due to the assumption.

Bug: 281452823
Test: atest VirtualSensorTest
Change-Id: Ifccf6d2f6f65140b8c8e27d7bd8add501a54fd2d