Explicitly skip the test for automotive devices.
Automotive devices may use CarFrameworkPackageStubs instead of a real
DocumentsUi to pass CTS. When it is used, we may get a resolved activity
for DocumentsUi, but it does not actually exist. The test will fail in
this condition. We can just skip this test as it does not mean anything
on automotive devices.

In addition, added a NPE check because PackageManager.resolveActivity()
may return null.

Bug: 260801840
Test: atest-src CtsAppTestCases:DownloadManagerTest#testDownload_onMediaStoreDownloadsDeleted
Change-Id: Ibd19d18c2da06f09ac325370707649ecc4c6eda8