Fix waiting for debuggable apps to attach
Previously, the test would explicitly wait for the debuggable app
activity start to complete with the activity-start launch flag -W.
This would never return because the application attach would be blocked
on waiting for a debugger connection.

Now, we start the activity asynchronously since we already spin loop to
get the pid and connect to the debugger.

Test: atest JdwpTunnelTest
Bug: 253908737
Fixes: 288959213
Change-Id: I97c1d984c47d6c4ff317c16c0100b95de059cbd8