[DO NOT MERGE] Allow pre-existing phantom process before testing
Improve robustness of testTrimAppChildProcess test case to handle initial condition on the system.

This patch is a fix to be able to run testTrimAppChildProcess with an initial count of phantom processes running as a test pre-condition.
The change is from "there is no existing phantom process" to "read count of existing phantom processes" and take it in consideration as a test pre-condition.

Bug: 286326104
Test: atest CtsAppTestCases
Change-Id: I38b37a0d97a15aa066683d6ab0aea732f383c652