Reduce drag time to make fling more likely during tests
Bug: 288339364

Two tests were executing a drag event and expecting a fling. The
drag lasted 2 seconds and if there was slowness in the test device,
it is possible that the fling was being interpreted as a normal
drag.

This CL drops the drag duration to 300ms so that it is more likely
that the drag becomes a fling.

Test: ran on forrest and locally
Change-Id: I04377da35dae2f2b167059b6c5adeb55efef94e0