Improve (un-flake) DocumentsClientTest
Improve on the logic for waiting for the root list to appear in the
DocumentsClientTest.revealRoot(...) method: instead of waiting for 30
seconds for the root list before checking if the drawer has to be open,
wait in 1 second intervals, and check for the presense of the drawer
opening "hamburger" button after each inteval.
This change significantly speeds up the test cases that start in the
state where the drawer is closed (by ~30sec per test case) as well as
"un-flakes" such test in case of the short lock screen timeout.

Bug: 287909283
Bug: 290309061
Test: atest CtsAppSecurityHostTestCases:android.appsecurity.cts.DocumentsTest
Change-Id: I51004272fa6da91fd740447b4710d77cc0cf3ed9