Reset DeviceConfig for display sandboxing CTS
Previously DeviceConfig for display sandboxing was not reset
before running the test. This change fixes this by explicitly
setting flags to null.

Also there is a small improvement in speed of the tests,
for some reason activity bounds were not reset to their
original values but to different values after the size of the
display gets restored. This problem is not important for
these tests to run because the Size Compat Mode gets disabled
as expected. Checking that the size changes is sufficient and
makes tests run faster.

Change-Id: I787ac6b6fda11c7a0c24c7cf4c2b4f901524394b
Test: atest CompatChangeTests
Bug: 289306840