[cts/pm] remove multi-user test from PackageSettingTest
The instrumenation of the device-side test app seems to be
interferencing with the permission state, but it is not reproducible
locally. Since the multi-user firstInstallTime test has been covered in
PackageManagerShellCommandMultiUserTest, it's okay to remove that part
from this test, so that the host-side test will only focus on
serialization/deserialization of PackageSetting during reboot.

BUG: 290087893
Test: atest com.android.tests.packagesetting.host.PackageSettingTest#testFirstInstallTimeWithReboot
Change-Id: I73d23c3b8dcfb7ca4d88ec17d1102060097dc12f