DebugInputRule: Use quotes only for empty setprop values
When setting the setprop value through the shell command, the value
was being wrapped in double-quotes. If the value was non-empty, the
quotes ended up being stored as part of the stored sysprop. Thus,
every time the previous value was being restored, it was nested in
another layer of quotes, resulting in the sysprop value growing with
each iteration (e.g. DEBUG -> "DEBUG" -> ""DEBUG"", etc.).

Additionally, because the setprop command expects a non-empty value
argument, it is not possible to un-set a property or set it to an empty
value.

To get around this, set the log tag property to UNKNOWN when
attempting to un-set the sysprop value (i.e. if it was not set or empty
before the test run). If restoring a non-empty sysprop value, do not
wrap it in quotes.

Bug: 290927121
Change-Id: Id79ed390f83a74e4bc982ce7038cf3d79704790c
Test: atest StylusButtonInputEventTest --rerun-until-failure