[cts] Move long AppEnumeration tests to post submit
Each of these test takes 3~30 seconds to make sure that a broadcast is not received. Some include the waiting time of waitForBroadcastBarrier() which can sometimes take a minute if the system is busy. Moving these tests to postsubmit so that they don't push the test suite out of SLO.

FIXES: 290941181
BUG: 288461477
Bug: 302750251
Test: presubmit

Merged-in: I965d0f459bf9b3c21b5ad4cde134493b6b71b7ea
Change-Id: I965d0f459bf9b3c21b5ad4cde134493b6b71b7ea