Disable OnBackInvoked for CaptureEventActivity
BackKeyShortcutsTest verifies the various key combinations generate a
KEYCODE_BACK key event. On some configurations the back key will trigger
an `OnBackInvokedCallback` that consumes the key event and causes the
test to fail. In this CL we explicitly configure the
`CaptureEventActivity` to disable these callbacks so that we can test
that the KEYCODE_BACK events are generated.

Bug: 287393717
Test: atest CtsInputTestCases:android.input.cts.BackKeyShortcutsTest
Change-Id: Id894b8309431f9b3bfd15db9c34c9e2045df774e