Extend IncompleteMotionTest timeout
While looking for flakes, it looks like there's a particular
timeout in IncompleteMotionTest that is failing occasionally.
I could not reproduce the flake locally, and from the logs, my
impression is that this is probably linked with the system being
under pressure. The timeout that I identified is for multiple
operations that will perform a 1 second sleep; since the timeout
is 3 seconds, that will make the test fail if the secondary
activity takes too long to run.

Bug: 288321659
Test: `atest StylusButtonInputEventTest`
Change-Id: I84fde452dabec534fefbe0a34756d5aaccac8b36