Correct instructions for device owner disallow add user test
The instructions ask the user to click the add user button in Settings,
to confirm that the policy transparency dialog is shown. But when
the restriction is set, it's no longer possible to access that setting,
instead the dialog is shown when the user tries to toggle "allow
multiple users".

Test: device owner > policy transparency > disallow add new user
Bug: 290811623
Change-Id: I923216ae0d251c9294e6982f8e4087e3a7cf4129