Don’t launch task in fullscreen in non-split cases
This CL is a partial revert of
ag/I6536c497eecb192a8f95cd9d7766c6e8cb0404a4.
The test launches the task on the secondary display in fullscreen to put
it in splitscreen launch root after the display disconnection, but that
should be only for splitscreen cases.
For non-splitscreen cases, the task can inherit the display windowing
mode, and continue to do it after display’s disconnection.

Bug: 290567729
Test: atest MultiDisplayPolicyTests (on cf_phone, cf_pc, and ARC)
Change-Id: I6505f3af7344ffb5147c932f9f9cdafd2c24d87b