[cts] remove flaky slowUiThreadWithKeyEventTriggersAnr test
The test is supposed to verify that an app will ANR if it
is doing too much work on the UI thread. The test is flaky
because the test may try to inject a key before the app has
consumed the focus event. Input dispatcher will wait and
then drop the key since there is no focused window.

To deflake this we can explictly wait for the app to consume
the focus event before slowing down the UI thread. A simpler
way would be to slow down the UI thread by delaying the
onKeyEvent or onTouchEvent callbacks - which is exactly what
slowOnKeyEventHandleTriggersAnr and slowOnTouchEventHandleTriggersAnr
test does. So we just remove the flaky test.

Fixes: 293699384
Test: atest AnrTests
Change-Id: I15f11237020a3c6175ba9ff44662b4825355bf68