Attempt to fix flaky testStructuredPhoneticName. Details: go/phoneticnameflakiness
Log files showed log messages stating that added test accounts have been changed and deleted. The flaky test doesn't trigger any such call but the test that ran before it in both the cases adds test sim accounts to assert correct addition and retrieval of sim accounts and then deletes them in its tearDown(). Deep dive into the code flow pointed out that this process might still be running when our flaky test started executing and so the test account added by it was marked as deletable(more details in the doc) which led to deletion of test raw contacts associated with it which in turn leads to flaky behaviour. By changing the flaky test's added test account to a sim account, we ensure that it is not marked as deletable without effecting our test logic in any way.

Bug: 286618555

Test: Flaky behaviour could not be reproduced locally. But the changes
have been tested locally. Test to be observed over the next few weeks to
verify it is no longer flaky.

Change-Id: I864da591131f369d9ebe91dde6ca54235c665ffd