Remove unecessary checks from PowerPolicyHostTest
OEMs can implement own policies for ON state, as result,
checks for system_power_policy_all_on will fail. Check for
system_power_policy_all_on was removed from test.
Test was verifying if after restart power policy is the same as at
the start of the test. This expectation might be incorrect if OEM
use diferrent power policies in ON state. This check was also removed.

Bug: 293892322
Test: atest android.car.cts.PowerPolicyHostTest
Change-Id: I62a550a98191fab11c041d22b1c80becf208180b