Rotate device from the current rotation
The test, testSplitAttributesCalculatorInvocation_screenRotation, failed
because we assume the device always initialize with ROTATION_0.

As the result, the verification fails if the device is at the rotation
state other than ROTATION_0 since the device state stayed unchanged
and didn't trigger the calculator callback.

This CL changes to rotate the device three times from the current
rotation to relax the assumption that device is always at ROTATION_0
before the test.

Test: atest
CtsWindowManagerJetpackTestCases:SplitAttributesCalculatorTest
--iterations
Fixes: 293935642

Change-Id: I5301c83bed32cf4d66b96cdebf66b46c0abe1816