Fix testDisplayShapeConfig
The test should always pass if the display doesn't have a rounded
corner. But it would fail if the display shape is configured to be
non-rectangular.

There was also a bug that the test didn't iterate all 4 rounded corners,
but only POSITION_TOP_LEFT.

This CL fixes the bugs and assumes the device has a default display
before running the test.

Fix: 292477517
Fix: 293342221
Test: atest DisplayShapeTests
Change-Id: I374f497d7a521721eff7726ee5b6e4337bccb9bd