CTSTelecomTestCases test improvment
This CL performs a cleanup on the EmergencyCallOnSimCallManagerTest
tests. The tests fail to call setUp() on
CtsSimCallManagerConnectionService, which causes the calls to fall back
on the CtsConnectionService.

The fix is to ensure we set up the connection service for
CtsSimCallManagerConnectionService instead of CtsConnectionService.
Another issue was that the emergency account created in that test was
causing the original phone account to be overwritten
(TEST_SIM_EMERGENCY_PHONE_ACCOUNT_HANDLE). We need to create another
phone account handle for emergency use cases with a separate id.

BaseTelecomTestWithMockServices required a small refactor to ensure that
the base functions know which connection service to reference in the
assertions.

Bug: 291705037
Test: atest CtsTelecomTestCases
Change-Id: Ic32b0826afe5157bb58177bbafb833a82b7f7965