Allow non-handheld device to skip Audio Loopback Latency test.
Examples of non-handheld devices are Android TV, Wear, Auto, etc.
They have difficulty running the loopback test so
we give them an automatic pass.

This is a manually back-ported version of aosp/2679835.
That was a trimmed down version of ag/24117501.

Bug: 260765871
Bug: 286376836
Test: Run Audio Loopback Test test on handheld. Test should be required.
Test: Run Audio Loopback test on non handheld (TV, watch, Auto).
Test:   Should skip test and automatically pass.

Change-Id: I2b092ebb503bbc06dd33fbdcef2fe629aee515d3
Merged-In: I32e6e7933b20b06fe23bc7400777646d4d071139