Update PowerPolicyHost test
Replaced CPMS On state check with wait for ON state.
Depending on implementation, if device is restarted before test,
it could be in WAIT_FOR_VHAL state instead of ON when becomes
available over adb. This was leading to flaky test results.

Test: atest CtsCarHostTestCases
Bug: 295212266
Bug: 296050172
Change-Id: Iaf98c191ae4d48f1f91dbed161c6059891c01e87
Merged-In: Ib83bdfa17448f0f8b2776d36d45d243ccce02865