Refactorings for future tests
Introduce a client for TestService to allow overloads and changes in
TestService signature without having to modify every call. This reduces
a lot of extra null/empty/default values.

Also introduces a way to start PendingIntents from the foreground
activity (similar to the ability of starting Intents). And to optionally
finish the activities started by those PendingIntents.

Test: atest BackgroundActivityLaunchTest
Change-Id: I9415410539c958ddc24310a7c310ccfdddabc18b