Grab data dependencies for CtsBionicTestCases from bionic.
This diverged and broke some memtag tests.

Yes, this probably will need to be patched up when the UDC push happens.
I'm assuming that bionic will be left as-is, and CTS will be clobbered.
But hopefully that'll be caught in the merge by someone and they'll let
me know (otherwise, test-suites won't build).

Bug: 296832150
Test: atest bionic-unit-tests CtsBionicTestCases
Change-Id: I3b462119bd5789808803aa2646f75aa7b3e5039f