Relax position assertion in testSetPlaybackParamsZeroSpeed
A specific order of audio and video buffer queueing can result in
MediaPlayer.getCurrentPosition() returning the timestamp of the second
audio buffer when playback is paused at the start. This is fixed by
http://r.android.com/2700283, but this change relaxes the CTS assertions
for older builds (that don't include the fix) to ensure they don't see
flaky failures as a result.

RESTRICT AUTOMERGE

Test: atest MediaPlayerTest
Bug: 283882552
Bug: 282472894
Bug: 263063118
Change-Id: I4713568e6c6ac0c5f92bc6cc50fa9e4bc5c38dc3