[RESTRICT AUTOMERGE] Fix CTS KeyAttestationTest
To fix the failures from android.keystore.cts.KeyAttestationTest#testEcAttestation

Check unexpected OID extension if the device launching with api_level>=30.
The test case checks if there's any unexpected extension of the attestation certificate,
while it detected OID: 2.5.29.35 (for authority key identifier) and OID: 2.5.29.15 (for Key Usage)
This unexpected certificate extension check is introduced on Android S (or R) but not on Android Q;
On Android Q, the TEE DOES add these extensions to the attestation certificate, while on Android S,
the reference TEE implementation doesn't add these extensions to the attestation certificate.

Bug: 294468310
Test: CtsKeystoreTestCases -t android.keystore.cts.KeyAttestationTest#testEcAttestation

Change-Id: Ic4bac68e6052d8cdd3c21b895baa785562346b6c
Signed-off-by: SeungJun, Jay Lee <seungjun.lee@gm.com>