Relax position assertion in testSetPlaybackParamsZeroSpeed
A specific order of audio and video buffer queueing can result in
MediaPlayer.getCurrentPosition() returning the timestamp of the second
audio buffer when playback is paused at the start. This is fixed by
http://r.android.com/2700283, but this change relaxes the CTS assertions
for older builds (that don't include the fix) to ensure they don't see
flaky failures as a result.

This is a manual 'cherry-pick' of ag/24514888 because the file has
moved between android-12L-tests-dev and android-13-tests-dev, so I made
the equivalent change again manually from scratch. This change should
auto-merge through from this branch (android-11-tests-dev) through to
android-12L-tests-dev and no further (because it will see that the
change ID referenced below is merged in android-13-tests-dev via
http://r.android.com/2726917).

Test: atest MediaPlayerTest
Bug: 283882552
Bug: 282472894
Bug: 263063118
Merged-In: I4713568e6c6ac0c5f92bc6cc50fa9e4bc5c38dc3
Change-Id: I32e3228467f974487036de28678a7d2e2b0fd2c0