Wait for idle instead in #testParentWindowMetricsPredicate
In the test, primary activity's maxBounds may be stale
because the updated config have not been applied even
after resume state.
A possible scenario is that a new config is coming and
the activity drop the previous config and wait for applying
the newer config.

This CL changes to wait for ActivityThread idle to verify
the split bounds.

Test: atest
ActivityEmbeddingBoundsTests#testParentWindowMetricsPredicate
--iterations on cf_x86_64_tablet-trunk_staging-userdebug
Bug: 293266079

Change-Id: Ie83bc7d62f7b3904db30922f279f22028fa2b718