Don't treat transparent pixels as unredacted pixels
AAOS can render activies in a non-full screen task view. The area of the
display not part of the task view shows up transparent when taking a
screen shot of the activity in the task view. These pixels shouldn't
count as unredacted.

Fixes: 297120391
Test: atest CtsDevicePolicyTestCases
Merged-In: I7181551f366fe5cf65841b96c8483517bce9b15f

Change-Id: I7181551f366fe5cf65841b96c8483517bce9b15f