Fix flaky ViewTreeObserverTest#testRemoveOnPreDrawlistener
Fixes: 294066671

Fixes: 299547089 (for cherry-pick)

There was a chance that a regular draw call could happen
in the middle of the test, causing the test to fail.
testRemoveOnPreDarawListener() was operating on the test
thread, so the UI thread was free to run during the test.

This CL moves the important part of the test to the UI
thread so that no draw calls can be done during the
test.

Test: ran test 100 times on forrest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:411536d28fdcce07ce512f9b51e4d70fb69b2cdd)
Merged-In: I350b77a1d42014689665aaa311904cbdd872b235
Change-Id: I350b77a1d42014689665aaa311904cbdd872b235