Switch PkgInstallSignatureVerificationTest to JUnit4
The test was written when BaseAppSecurityTest did not exist. Switch to
BaseAppSecurityTest as the majoirty of the classes in the same package
use. Some other tests are using DeviceTestCase and but is unrelated to
this change.

The parent class change implies inheriting from BaseHostJUnit4Test.
Adjust the annotation accordingly for JUnit4.

Bug: 277347456
Test: atest android.appsecurity.cts.PkgInstallSignatureVerificationTest
      # 132 tests passed; the number is same without this change.
Change-Id: I53ce4e064aae674a1ae34ac14b7ed002ecabb0c5