ctsv: prevent ANRs in RingerMode test
The Volume test was taking more than 5 seconds
and was running in the UI thread. This could cause
an ANR if the screen was touched.

This fix breaks the test into 3 parts and reduces some of the sleep
times.

It also logs the test duration and logs a warning if over 5 seconds.

Bug: 271171384
Test: Run CTS Verifier "Ringer Mode Tests"
Test: While running, scroll screen up and down.
Change-Id: I44ea39463982a8929d41eab50f53ca81b5b931cf