Test APK install with v4 signature off incremental
Parameterize the existing tests of incremental install to also test the
case without incremental. Most existing test expectations remain the
same.

Some tests do require adjustment. fs-verity does not use the Merkle tree
in .idsig, so if the tree is broken, it does not prevent the install
unlike incremental. The tree is simply ignored. The tests are:
 * testInstallV4WithNoMerkleTree
 * testInstallV4WithMerkleTreeBitsFlipped
 * testInstallV4WithWithTrailingDataInMerkleTree
 * testInstallV4With10MBytesAdditionalDataFails
 * testInstallV4With256BytesAdditionalDataFails

While one of the test is expected to pass, apparently an idsig is
problematic: v4-digest-v3-merkle-tree-1mb-trailing-data.apk.idsig. The
file contains a wrong raw root hash and signature. This change fixes
that. It wasn't a problem previously since the test is supposed to be
fail anyway.

Bug: 277347456
Test: atest CtsAppSecurityHostTestCases
Change-Id: Ie600a75b040c050313d0ae854ac8a280c9152487