Camera: Simplify #testCreateSessionWithParameters
'#testCreateSessionWithParameters' can in some cases try to
configure a reprocessable stream combination that is not
part of the mandatory table.
Since this test case intends to test the session parameter
configuration using a reprocessing session, simplify the
stream combination to ensure that it will be supported
on all devices advertising the private and/or YUV reprocessing
capability.

Bug: 300661768
Test: atest android.hardware.camera2.cts.CameraDeviceTest#testCreateSessionWithParameters

Change-Id: I1963770dcafda3b80794aa119219983e7b2d2338