MediaRecorderTest: Increase file size for 720p cameras
The ag/10876653 introduced the CamcorderProfile for getting the video
size. The camera's LOW_QUALITY profile may be as high as 720p. Depending
on the type of the camera, the single encoded frame may be already
exceeding the max file size, which currently equals to 50k.

It's also possible that the camera could return the minimum supported
resolution to be 720p, which would lead to a similar failure.

Increasing the max file size to 250kB for 720p videos should be enough
to make the testRecorderVideo() and testSetOutputFile() tests pass for
the problematic devices.

Bug: 298743195
Test: CtsMediaRecorderTestCases android.media.recorder.cts.MediaRecorderTest#testSetOutputFile
Test: CtsMediaRecorderTestCases android.media.recorder.cts.MediaRecorderTest#testRecorderVideo
(cherry picked from https://android-review.googlesource.com/q/commit:19074125a7a27ba1347924da5925377242fe2e94)
Merged-In: I3c9402dce29030098974353fef1272c7f6c34437
Change-Id: I3c9402dce29030098974353fef1272c7f6c34437