Simplify testAppRelaunchSetsWaitResultDelayData
There was a race that the test activity is not stopped yet because
transition is still running. Then if it is launched again, it may
commit the previous pending draw state to report drawn. That causes
reportActivityLaunched to finish too early before finishRelaunching
is reached.

Because the purpose of the test is to bring an existing activity
from background to foreground with configuration change, now the
test simply returns to home and wait for it to be stopped. So the
test activity won't keep the pending draw state.

Bug: 301193935
Test: ActivityMetricsLoggerTests#testAppRelaunchSetsWaitResultDelayData
Merged-In: I610dbb6bb95cb8a9026582992c41b47f645b09f9
Change-Id: I610dbb6bb95cb8a9026582992c41b47f645b09f9
(cherry picked from commit 414f9bf1c861c7175a09fa9e4c8766df8a40f1ab)