Directly launch activity in secondary split root
This CL changes the target activity to be launched directly in the
secondary split root instead of launching in fullscreen and then moving
it to split mode.
The test case verifies that the app-requested orientation is not
respected in splitscreen. However, because currently the portrait
activity is launched in fullscreen first and then moved to splitscreen,
the display rotates when fullscreen and rotates back when split mode,
which causes the flakiness of the test.
Also, launching an activity in fullscreen may deactivate the splitscreen
mode in some devices (e.g., ARC++).

Before TestTaskOrg was enabled, we used to launch it directly in the
secondary split root. So this CL just restores the original behavior.

Bug: 299574293
Test: atest android.server.wm.AppConfigurationTests (cf_phone and ARC)
Change-Id: I727bc705696e47a447f3b1599c1e79bfc8838f4b