Use unbuffered dispatch in VirtualTouchscreenTest
Inside VirtualTouchscreenTest#sendTouchEvent, we are injecting
ACTION_DOWN, ACTION_MOVE, and ACTION_UP events into the virtual display.

The ACTION_MOVE events occasionally get resampled. This leads to the
modification of their coordinates.

The resampling is not exposed to apps, so we can't test it easily. Since
we are asserting the receipt of exact coordinates that were injected,
let's use unbuffered dispatch in this test.

Bug: 292544797
Test: atest VirtualTouchscreenTest --iterations 20
Change-Id: I510a0d516a1f0a513265debf8f32f47674ef2702