Fix virtual display rotation test for tablets.
Tablets set config_reverseDefaultRotation to true, which results in
portrait orientation of 90 instead of 270.

This is a short term fix, we should consider whether virtual displays
should respect the same config, or whether to reverse the default
rotation for virtual displays automatically based on whether they have
portrait or landscape dimensions in ROTATION_0.

Fix: 302121290
Test: presubmit
Change-Id: I3b8f25405149ba54f0693e121be341f4fbf1bf65