[DO NOT MERGE] Workaround the testWindowProducerScaling CTS issue
Slightly increase the size of the test sample in `testWindowProducerScaling`
to avoid the test failure caused by the "rounding error", when the status
bar height overridden, conbined with
`config_remoteInsetsControllerControlsSystemBars = true`.

For a proper fix the test needs to be slightly reworked to not capture
areas covered by system bars.

Bug: 300268614
Test: atest-src android.view.cts.PixelCopyTest -c
Change-Id: I9f5c97f75208604796eb12e1086c56f30b91f8ab