Verify callback is invoked regardless of success
The query may fail for various reasons including a timeout or failure connecting to the Geocoder backend, causing the test to be horrendously flaky. Instead, the test should purely excercise the API surface and not the backend. Hence with this change the test will verify that the callback is invoked with success or error regardless of the success of the query to the backend.

This change also removes the annotation that currently ignores this test in presubmit.

The timeout for `verify` is longer than the GRPC timeout so the test should still pass even if the request timesout.

Test: atest --iterations=100 /usr/local/google/home/sashakuznetsov/src/android/main/cts/tests/location/location_fine/src/android/location/cts/fine/GeocoderTest.java

Bug: 301220358

Change-Id: I4ee5d16b8ae08a3427103efd212fb091a22ea541