Remove testLightBarIsNotAllowed_fitDisplayCutout
Assume a display cutout is at the same side of status bar and provides
less insets. When a window fits display cutouts, the window would cut
halfway into status bar, and status bar background would be semi-
transparent. The color of status bar would be unpredictable. The
behavior is allowed.

testLightBarIsNotAllowed_fitDisplayCutout was ignored previously because
it cannot handle the case mentioned above. This CL removes the test.

Bug: 292476663
Test: atest LightBarTests
Change-Id: Idfe9f0b6af08f96fccda76891af7221a50be64a1