Camera: Fix the failing cts test issue
Currently the CaptureResultTests are failing while trying to check
whether the manual flash control flag is enabled or not. This flag value
cannot be read from Flags class during runtime due to a known issue.
Since it is blocking CTS, commenting out the block of code which reads
this flag value until the b/301272559 is resolved.

Bug: 303468894

Test: Tested locally on felix
Change-Id: Ia8260f02dba52e64f8fec73685690499425f31b5