Fix AccessibilityOverlayTest failure when modify logical display sizes
The width and height of the test window is 1x1. After wm size scaling, when calculating mTouchableRegionInWindow and mTouchableRegionInWindow of a11yWindow, due to the scaling ratio of displayMatrix, the calculated width and height are empty:
regionInScreen=SkRegion(), touchableRegion=SkRegion(), which in turn causes the unaccountedSpace.quickReject(regionInScreen) condition in the judgment of windowMattersToAccessibility() to be unsatisfied and not added to Accessibility.

Bug: 301218644
Test: CtsAccessibilityServiceTestCases.android.accessibilityservice.cts.AccessibilityOverlayTest#testA11yServiceShowsDisplayOverlayUsingSurfaceControl_shouldAppearAndDisappear
Merged-In: I04b5f474ed4e53c43f9486034d73700a05d2208a
Change-Id: I04b5f474ed4e53c43f9486034d73700a05d2208a