Fix flaky AudioFocusTest#testAudioFocusRequestGainLoss*
The calls to onAudioFocusChange() and waitForFocusChange()
are synchronized through a lock. In addition a semaphore
is used to handle the wait for the focus change callback.
As these methods can be called from different threads, it
can lead to update of AudioFocus state interrupted before
the semapore permit is released.

This impacts the subsequent calls of requesting or abandoning
audio focus. Instead of waiting for the new semaphore permit,
the call immediately returns althrough the audio focus state
is not received yet. Note, the callback indeed is received
subsequently and in-time.

Fix this by making sure the waitForFocusChange() does not
return until the permit is available. This way, the permit
is accounted for the focus change and not just the focus
state.

Bug: 303223553
Bug: 302870089
Test: atest -c --rerun-until-failure 1000 AudioFocusTest

Change-Id: I7bc55c84bdd86237032b1b5430605a58e9847257